sync fwk

+++ engine/split/v4k_render.h
@@ -15,7 +15,7 @@ API unsigned rgba( uint8_t r, uint8_t g, uint8_t b, uint8_t a );
 API unsigned bgra( uint8_t b, uint8_t g, uint8_t r, uint8_t a );
 API unsigned rgbaf( float r, float g, float b, float a );
 API unsigned bgraf( float b, float g, float r, float a );
-API float    alpha( unsigned rgba );
+API unsigned alpha( unsigned rgba );
 
 #define RGBX(rgb,x)   ( ((rgb)&0xFFFFFF) | (((unsigned)(x))<<24) )
 #define RGB3(r,g,b)   ( (255<<24) | ((r)<<16) | ((g)<<8) | (b) )
